// src/features/base/discover/types.ts
export type Attraction = {
  id: string;
  image: string;
  category: string;
  title: string;
  description: string;
};

export type DiscoverCard = {
  id: string;
  title: string;
  description: string;
};

export type TimelineDay = {
  id: string;
  day: string;
  title: string;
  description: string;
};